+++ src/main/java/subway/domain/Station.java
@@ -1,4 +1,4 @@
-package subway;
+package subway.domain;
 
 import javax.persistence.*;
