+++ src/main/java/subway/domain/StationRepository.java
@@ -1,4 +1,4 @@
-package subway;
+package subway.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
